+++ home.py
@@ -4,7 +4,7 @@ conference_page = st.Page(
     "conference_app/app.py", title="Book Conference", icon="🏢", url_path="conference"
 )
 resource_page = st.Page(
-    "resource_app/app.py", title="Request Resource", icon="🤖", url_path="resource"
+    "resource_app/app.py", title="Request Resource", icon="🛠️", url_path="resource"
 )
 
 pg = st.navigation([conference_page, resource_page], position="top")